requirements.txt
langchain==1.0.4
langchain_community==0.4.1
langchain_google_genai==3.0.1
langchain_huggingface==1.0.1
python-dotenv==1.2.1
